+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * Remove `super.componentDidMount()` and `super.componentDidUpdate( prevProps, prevState )` calls from
   the `ReplicantSubscription` react component as not needed as parent methods are empty as of react4j
   version `0.96`
+* Fix bug where converger will not re-converge if multiple AreaOfInterest are added simultaneously and the
+  the later `AreaOfInterest` instances can not be grouped into the first `AreaOfInterest` instance. The
+  converger would previously incorrectly halt after the first action completed.
 
 ### [v6.10](https://github.com/realityforge/replicant/tree/v6.10) (2018-09-21)
 [Full Changelog](https://github.com/realityforge/replicant/compare/v6.09...v6.10)